+++ build.gradle
@@ -36,6 +36,7 @@ dependencies {
     implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter:3.0.3'
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
     implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
+    testImplementation 'org.springframework.cloud:spring-cloud-starter-contract-stub-runner'
 
 
     compileOnly 'org.projectlombok:lombok'
