+++ src/test/java/com/kb/zipkim/ZipkimApplicationTests.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest(classes = ZipkimApplication.class)
 class ZipkimApplicationTests {
 
     @Test
